Draft: Reduce ScrollView content margins
Do not set extra margins that add up to the plasmoid margins.

Set them though on top, to have a separation from the header
and on the right when scrollbar is visible.

WIP since to match the plasmoid margins, we should extract them
from the plasma theme. Should this be done in a more general way
instead, maybe in PlasmaExtras.Representation, or add some helper
unit in `Plasma.Units` or `Plasma.Theme`?